Renamed homeView to QuizView
