+++ src/App.vue
@@ -4,6 +4,7 @@
       <RouterLink to="/">Home</RouterLink>
       <RouterLink to="/language">Language</RouterLink>
       <RouterLink to="/categories">Category</RouterLink>
+      <RouterLink to="/quiz">Quiz</RouterLink>
     </nav>
   </header>
   <main id="viewport">
