+++ src/router/index.js
@@ -1,5 +1,5 @@
 import { createRouter, createWebHistory } from "vue-router";
-import HomeView from "../views/HomeView.vue";
+import QuizView from "../views/QuizView.vue";
 import LanguageView from "../views/LanguageView.vue";
 import LandingView from "../views/LandingView.vue";
 import CategoryView from "../views/CategoryView.vue";
@@ -16,7 +16,7 @@ const router = createRouter({
     {
       path: "/quiz",
       name: "quiz",
-      component: HomeView,
+      component: QuizView,
     },
     {
       path: "/language",
